+++ docs/changelog.md
@@ -23,6 +23,9 @@ title: "Changelog"
   - `tslib` to 2.0.1
   - `@types/mongoose` to 5.7.36
 - Allow any argument to `DeferredFunc`
+- Add Type `DynamicStringFunc`
+- Allow definition of functions for `localField` & `foreignField`
+- Allow returning an function in `ref` (`ref: () => (doc) => doc.someProp`)
 - Allow the NestJS / Type-GraphQl way of defining arrays (`type: () => [String]` (and nested too))
 - [IC] Fix that `dim` is included as an option in the schema [[typegoose#366](https://github.com/typegoose/typegoose/issues/366)]